Refactored and moved the model porting code.
